chore: split classes for core and warp
