+++ typescript/radix-sdk/src/index.ts
@@ -4,14 +4,16 @@ import { utils } from 'ethers';
 
 import { assert, strip0x } from '@hyperlane-xyz/utils';
 
-import { RadixPopulate } from './modules/populate.js';
-import { RadixQuery } from './modules/query.js';
-import { RadixTx } from './modules/tx.js';
-import { Account, RadixSDKOptions } from './types.js';
-import { generateNewEd25519VirtualAccount } from './utils.js';
+import { RadixCorePopulate } from './core/populate.js';
+import { RadixCoreQuery } from './core/query.js';
+import { RadixCoreTx } from './core/tx.js';
+import { RadixBase } from './utils/base.js';
+import { Account, RadixSDKOptions } from './utils/types.js';
+import { generateNewEd25519VirtualAccount } from './utils/utils.js';
+import { RadixWarpPopulate } from './warp/populate.js';
+import { RadixWarpQuery } from './warp/query.js';
+import { RadixWarpTx } from './warp/tx.js';
 
-// TODO: RADIX
-// add mainnet package address after deploying it over https://console.radixdlt.com/deploy-package
 const NETWORKS = {
   [NetworkId.Stokenet]: {
     applicationName: 'hyperlane',
@@ -20,7 +22,8 @@ const NETWORKS = {
   },
   [NetworkId.Mainnet]: {
     applicationName: 'hyperlane',
-    packageAddress: '',
+    packageAddress:
+      'package_rdx1pk3ldj3ktxuw6sv5txspjt2a8s42c7xxcn6wnf5yuytdrcqhpflfkc',
   },
 };
 
@@ -35,8 +38,15 @@ export class RadixSDK {
   public applicationName: string;
   public packageAddress: string;
 
-  public query: RadixQuery;
-  public populate: RadixPopulate;
+  public base: RadixBase;
+  public query: {
+    core: RadixCoreQuery;
+    warp: RadixWarpQuery;
+  };
+  public populate: {
+    core: RadixCorePopulate;
+    warp: RadixWarpPopulate;
+  };
 
   constructor(options?: RadixSDKOptions) {
     this.networkId = options?.networkId ?? NetworkId.Mainnet;
@@ -54,13 +64,26 @@ export class RadixSDK {
       networkId: this.networkId,
     });
 
-    this.query = new RadixQuery(this.networkId, this.gateway);
-    this.populate = new RadixPopulate(
+    this.base = new RadixBase(
+      this.networkId,
       this.gateway,
-      this.query,
-      this.packageAddress,
       options?.gasMultiplier ?? DEFAULT_GAS_MULTIPLIER,
     );
+
+    this.query = {
+      core: new RadixCoreQuery(this.networkId, this.gateway, this.base),
+      warp: new RadixWarpQuery(this.networkId, this.gateway, this.base),
+    };
+
+    this.populate = {
+      core: new RadixCorePopulate(this.gateway, this.base, this.packageAddress),
+      warp: new RadixWarpPopulate(
+        this.gateway,
+        this.base,
+        this.query.warp,
+        this.packageAddress,
+      ),
+    };
   }
 
   public getNetworkId() {
@@ -71,19 +94,31 @@ export class RadixSDK {
 export class RadixSigningSDK extends RadixSDK {
   private account: Account;
 
-  public tx: RadixTx;
+  public tx: {
+    core: RadixCoreTx;
+    warp: RadixWarpTx;
+  };
 
   private constructor(account: Account, options?: RadixSDKOptions) {
     super(options);
 
     this.account = account;
-    this.tx = new RadixTx(
-      account,
-      this.query,
-      this.populate,
-      this.networkId,
-      this.gateway,
-    );
+    this.tx = {
+      core: new RadixCoreTx(
+        account,
+        this.base,
+        this.populate.core,
+        this.networkId,
+        this.gateway,
+      ),
+      warp: new RadixWarpTx(
+        account,
+        this.base,
+        this.populate.warp,
+        this.networkId,
+        this.gateway,
+      ),
+    };
   }
 
   public getAddress() {
