+++ typescript/radix-sdk/tsconfig.json
@@ -6,8 +6,11 @@
   },
   "include": [
     "./src/*.ts",
-    "src/modules/populate.ts",
-    "src/modules/query.ts",
-    "src/modules/tx.ts"
+    "src/core/populate.ts",
+    "src/core/query.ts",
+    "src/core/tx.ts",
+    "src/utils/signer.ts",
+    "src/utils/types.ts",
+    "src/utils/utils.ts"
   ]
 }
